refactor

+++ src/players/players.service.ts
@@ -6,8 +6,8 @@ import {
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
-import { createPlayerDto } from 'src/dto/create-player.dto';
-import { updatePlayerDto } from 'src/dto/update-player.dto';
+import { CreatePlayerDto } from 'src/dto/create-player.dto';
+import { UpdatePlayerDto } from 'src/dto/update-player.dto';
 import { Player } from 'src/interfaces/player.interface';
 
 @Injectable()
@@ -18,7 +18,7 @@ export class PlayersService {
 
   private readonly logger = new Logger(PlayersService.name);
 
-  async createPlayer(createPlayer: createPlayerDto) {
+  async createPlayer(createPlayer: CreatePlayerDto) {
     const { email } = createPlayer;
     const playerExists = await this.playerModel.findOne({ email }).exec();
 
@@ -30,7 +30,7 @@ export class PlayersService {
     return await createdPlayer.save();
   }
 
-  async updatePlayer(_id: string, updatePlayer: updatePlayerDto) {
+  async updatePlayer(_id: string, updatePlayer: UpdatePlayerDto) {
     const playerExists = await this.playerModel.findOne({ _id }).exec();
 
     if (playerExists) {
@@ -59,6 +59,12 @@ export class PlayersService {
   async deletePlayer(
     _id: string,
   ): Promise<{ acknowledged: boolean; deletedCount: number }> {
+    const player = await this.playerModel.findOne({ _id }).exec();
+    if (!player) {
+      throw new NotFoundException(
+        `Jogador não encontrado com o ID informado: ${_id}`,
+      );
+    }
     return await this.playerModel.deleteOne({ _id });
   }
 }
